+++ components/steps/index.ts
@@ -1,6 +1,4 @@
 export { Steps } from './Steps'
-export { ProgressIcon } from './ProgressIcon'
-export { PanelArrow } from './PanelArrow'
 export type {
   StepItem,
   StepStatus,